update naming for types
- add main.ts
