+++ src/types/index.ts
@@ -4,8 +4,8 @@ export interface MergeOptions {
     videoSegments: string[],
     basename: string,
     isVideoSegmentKept: string,
-    totalTime: number,
-    timeDiff: number
+    videoDuration: number,
+    elapsedTime: number
 }
 
 export interface FFmpegArguments {
@@ -16,4 +16,15 @@ export interface FFmpegArguments {
 
 export type FFmpegConfig = typeof FFMPEG_OPTIONS
 
-export type RemoveVideoSegmentArguments = Omit<MergeOptions, "totalTime" | "timeDiff">
+export type RemoveVideoSegmentArguments = Omit<MergeOptions, "videoDuration" | "elapsedTime">
+
+export interface ReadlineQA {
+    question: string,
+    answer: string,
+}
+
+export interface ReadlineCloseCallback {
+    close: () => void
+}
+
+export type ReadlineMergeCallback = (answer: string, mergeOptions: MergeOptions, readlineInterface: ReadlineCloseCallback) => void;
